Centralize Railway GraphQL ops
